refactor(login): remove feature and ui libraries
