+++ apps/console-e2e/project.json
@@ -1,5 +1,4 @@
 {
-  "root": "apps/console-e2e",
   "sourceRoot": "apps/console-e2e/src",
   "projectType": "application",
   "targets": {
