+++ libs/pages/login/.babelrc
