+++ libs/pages/login/README.md
@@ -1,0 +1,7 @@
+# pages-login
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test pages-login` to execute the unit tests via [Jest](https://jestjs.io).
