--- libs/pages/login/feature/README.md
@@ -1,7 +1,0 @@
-# pages-login-feature
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test pages-login-feature` to execute the unit tests via [Jest](https://jestjs.io).
