+++ libs/pages/login/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/page-login'
