+++ libs/pages/login/src/lib/page-login.spec.tsx
@@ -1,10 +1,10 @@
 import { render } from '__tests__/utils/setup-jest'
 
-import Login from './login'
+import PageLogin from './page-login'
 
-describe('Login', () => {
+describe('PageLogin', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(<Login />)
+    const { baseElement } = render(<PageLogin />)
     expect(baseElement).toBeTruthy()
   })
 })
