+++ libs/pages/login/src/lib/page-login.tsx
@@ -1,0 +1,7 @@
+import { Container } from '../lib/feature/container/container'
+
+export function PageLogin() {
+  return <Container />
+}
+
+export default PageLogin
