--- libs/pages/login/ui/README.md
@@ -1,7 +1,0 @@
-# pages-login-ui
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test pages-login-ui` to execute the unit tests via [Jest](https://jestjs.io).
