--- libs/pages/login/ui/jest.config.ts
@@ -1,10 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'pages-login-ui',
-  preset: '../../../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]sx?$': 'babel-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../../coverage/libs/pages/login/ui',
-}
