--- libs/pages/login/ui/project.json
@@ -1,22 +1,0 @@
-{
-  "sourceRoot": "libs/pages/login/ui/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/pages/login/ui/**/*.{ts,tsx,js,jsx}"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/pages/login/ui"],
-      "options": {
-        "jestConfig": "libs/pages/login/ui/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  }
-}
