--- libs/pages/login/ui/src/index.ts
@@ -1,3 +1,0 @@
-export * from './components/login-callback/login-callback'
-export * from './components/login/login'
-export * from './components/layout-login/layout-login'
