+++ libs/pages/logs/infra/feature/project.json
@@ -1,5 +1,4 @@
 {
-  "root": "libs/pages/logs/infra/feature",
   "sourceRoot": "libs/pages/logs/infra/feature/src",
   "projectType": "library",
   "tags": [],
