+++ package.json
@@ -20,10 +20,8 @@
     "@nrwl/eslint-plugin-nx": "14.3.1",
     "@nrwl/linter": "14.3.1",
     "@nrwl/nx-cloud": "14.1.0",
-    "@nrwl/react": "14.3.1",
     "@nrwl/storybook": "14.3.1",
     "@nrwl/tao": "13.8.5",
-    "@nrwl/web": "14.3.1",
     "@nrwl/workspace": "14.3.1",
     "@radix-ui/react-accordion": "^0.1.6",
     "@radix-ui/react-dialog": "^0.1.7",
@@ -77,6 +75,8 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-typescript": "^7.12.13",
     "@nrwl/jest": "14.3.1",
+    "@nrwl/react": "14.3.1",
+    "@nrwl/web": "14.3.1",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^8.0.4",
@@ -91,6 +91,8 @@
     "@types/faker": "^6.6.9",
     "@types/jest": "27.4.1",
     "@types/node": "17.0.41",
+    "@types/react": "18.0.12",
+    "@types/react-dom": "18.0.5",
     "@types/react-router-dom": "5.3.3",
     "@types/redux-mock-store": "^1.0.3",
     "@typescript-eslint/eslint-plugin": "5.24.0",
