+++ tsconfig.base.json
@@ -25,8 +25,7 @@
       "@console/pages/application/ui": ["libs/pages/application/ui/src/index.ts"],
       "@console/pages/environments/feature": ["libs/pages/environments/feature/src/index.ts"],
       "@console/pages/environments/ui": ["libs/pages/environments/ui/src/index.ts"],
-      "@console/pages/login/feature": ["libs/pages/login/feature/src/index.ts"],
-      "@console/pages/login/ui": ["libs/pages/login/ui/src/index.ts"],
+      "@console/pages/login": ["libs/pages/login/src/index.ts"],
       "@console/pages/logs/infra/feature": ["libs/pages/logs/infra/feature/src/index.ts"],
       "@console/pages/logs/infra/ui": ["libs/pages/logs/infra/ui/src/index.ts"],
       "@console/pages/onboarding/feature": ["libs/pages/onboarding/feature/src/index.ts"],
