+++ workspace.json
@@ -14,8 +14,7 @@
     "pages-application-ui": "libs/pages/application/ui",
     "pages-environments-feature": "libs/pages/environments/feature",
     "pages-environments-ui": "libs/pages/environments/ui",
-    "pages-login-feature": "libs/pages/login/feature",
-    "pages-login-ui": "libs/pages/login/ui",
+    "pages-login": "libs/pages/login",
     "pages-logs-infra-feature": "libs/pages/logs/infra/feature",
     "pages-logs-infra-ui": "libs/pages/logs/infra/ui",
     "pages-onboarding-feature": "libs/pages/onboarding/feature",
